+++ app/gui/src/project-view/components/GraphEditor/widgets/WidgetTableEditor/__tests__/tableNewArgument.test.ts
@@ -1,4 +1,5 @@
 import {
+  CELLS_LIMIT,
   DEFAULT_COLUMN_PREFIX,
   NEW_COLUMN_ID,
   ROW_INDEX_HEADER,
@@ -7,7 +8,7 @@ import {
   useTableNewArgument,
 } from '@/components/GraphEditor/widgets/WidgetTableEditor/tableNewArgument'
 import { MenuItem } from '@/components/shared/AgGridTableView.vue'
-import { WidgetInput } from '@/providers/widgetRegistry'
+import { WidgetInput, WidgetUpdate } from '@/providers/widgetRegistry'
 import { SuggestionDb } from '@/stores/suggestionDatabase'
 import { makeType } from '@/stores/suggestionDatabase/entry'
 import { assert } from '@/util/assert'
@@ -21,8 +22,16 @@ function suggestionDbWithNothing() {
   return db
 }
 
+function generateTableOfOnes(rows: number, cols: number) {
+  const code = `Table.new [${[...Array(cols).keys()].map((i) => `['Column #${i}', [${Array(rows).fill('1').join(',')}]]`).join(',')}]`
+  return Ast.parse(code)
+}
+
 const expectedRowIndexColumnDef = { headerName: ROW_INDEX_HEADER }
 const expectedNewColumnDef = { cellStyle: { display: 'none' } }
+
+const CELLS_LIMIT_SQRT = Math.sqrt(CELLS_LIMIT)
+assert(CELLS_LIMIT_SQRT === Math.floor(CELLS_LIMIT_SQRT))
 
 test.each([
   {
@@ -111,6 +120,54 @@ test.each([
   expect(onUpdate).not.toHaveBeenCalled()
   expect(addMissingImports).not.toHaveBeenCalled()
 })
+
+test.each([
+  {
+    rows: Math.floor(CELLS_LIMIT / 2) + 1,
+    cols: 1,
+    expectNewRowEnabled: true,
+    expectNewColEnabled: false,
+  },
+  {
+    rows: 1,
+    cols: Math.floor(CELLS_LIMIT / 2) + 1,
+    expectNewRowEnabled: false,
+    expectNewColEnabled: true,
+  },
+  {
+    rows: 1,
+    cols: CELLS_LIMIT,
+    expectNewRowEnabled: false,
+    expectNewColEnabled: false,
+  },
+  {
+    rows: CELLS_LIMIT,
+    cols: 1,
+    expectNewRowEnabled: false,
+    expectNewColEnabled: false,
+  },
+  {
+    rows: CELLS_LIMIT_SQRT,
+    cols: CELLS_LIMIT_SQRT,
+    expectNewRowEnabled: false,
+    expectNewColEnabled: false,
+  },
+])(
+  'Allowed actions in table near limit (rows: $rows, cols: $cols)',
+  ({ rows, cols, expectNewRowEnabled, expectNewColEnabled }) => {
+    const input = WidgetInput.FromAst(generateTableOfOnes(rows, cols))
+    const tableNewArgs = useTableNewArgument(
+      input,
+      { startEdit: vi.fn(), addMissingImports: vi.fn() },
+      suggestionDbWithNothing(),
+      vi.fn(),
+    )
+    expect(tableNewArgs.rowData.value.length).toBe(rows + (expectNewRowEnabled ? 1 : 0))
+    const lastColDef = tableNewArgs.columnDefs.value[tableNewArgs.columnDefs.value.length - 1]
+    assert(lastColDef?.headerComponentParams?.type === 'newColumn')
+    expect(lastColDef.headerComponentParams.enabled ?? true).toBe(expectNewColEnabled)
+  },
+)
 
 test.each([
   'Table.new 14',
@@ -517,3 +574,35 @@ test.each([
     else expect(addMissingImports).not.toHaveBeenCalled()
   },
 )
+
+test('Pasted data which would exceed cells limit is truncated', () => {
+  const initialRows = CELLS_LIMIT_SQRT - 2
+  const initialCols = CELLS_LIMIT_SQRT - 1
+  const ast = generateTableOfOnes(initialRows, initialCols)
+  const input = WidgetInput.FromAst(ast)
+  const startEdit = vi.fn(() => ast.module.edit())
+  const onUpdate = vi.fn((update) => {
+    const inputAst = update.edit!.getVersion(ast)
+    // We expect the table to be fully extended, so the number of cells (numbers or Nothings) should be equal to the limit.
+    let cellCount = 0
+    inputAst.visitRecursive((ast: Ast.Ast | Ast.Token) => {
+      if (ast instanceof Ast.Token) return
+      if (ast instanceof Ast.NumericLiteral || ast.code() === 'Nothing') cellCount++
+    })
+    expect(cellCount).toBe(CELLS_LIMIT)
+  })
+  const addMissingImports = vi.fn()
+  const tableNewArgs = useTableNewArgument(
+    input,
+    { startEdit, addMissingImports },
+    suggestionDbWithNothing(),
+    onUpdate,
+  )
+  const focusedCol = tableNewArgs.columnDefs.value[initialCols - 2]
+  assert(focusedCol?.colId != null)
+  tableNewArgs.pasteFromClipboard(Array(4).fill(Array(4).fill('2')), {
+    rowIndex: initialRows - 2,
+    colId: focusedCol.colId,
+  })
+  expect(onUpdate).toHaveBeenCalledOnce()
+})