+++ app/gui/src/project-view/components/GraphEditor/widgets/WidgetTableEditor/tableNewArgument.ts
@@ -23,6 +23,11 @@ export const ROW_INDEX_HEADER = '#'
 export const DEFAULT_COLUMN_PREFIX = 'Column #'
 const NOTHING_PATH = 'Standard.Base.Nothing.Nothing' as QualifiedName
 const NOTHING_NAME = qnLastSegment(NOTHING_PATH)
+/**
+ * The cells limit of the table; any modification which would exceed this limt should be
+ * disallowed in UI
+ */
+export const CELLS_LIMIT = 256
 
 export type RowData = {
   index: number
@@ -180,10 +185,28 @@ export function useTableNewArgument(
     }
   }
 
+  function mayAddNewRow(
+    rowCount_: number = rowCount.value,
+    colCount: number = columns.value.length,
+  ): boolean {
+    return (rowCount_ + 1) * colCount <= CELLS_LIMIT
+  }
+
+  function mayAddNewColumn(
+    rowCount_: number = rowCount.value,
+    colCount: number = columns.value.length,
+  ): boolean {
+    return rowCount_ * (colCount + 1) <= CELLS_LIMIT
+  }
+
   function addRow(
     edit: Ast.MutableModule,
     valueGetter: (column: Ast.AstId, index: number) => unknown = () => null,
   ) {
+    if (!mayAddNewRow()) {
+      console.error(`Cannot add new row: the ${CELLS_LIMIT} limit of cells would be exceeded.`)
+      return
+    }
     for (const [index, column] of columns.value.entries()) {
       const editedCol = edit.getVersion(column.data)
       editedCol.push(convertWithImport(valueGetter(column.data.id, index), edit))
@@ -204,6 +227,10 @@ export function useTableNewArgument(
     size: number = rowCount.value,
     columns?: Ast.Vector,
   ) {
+    if (!mayAddNewColumn()) {
+      console.error(`Cannot add new column: the ${CELLS_LIMIT} limit of cells would be exceeded.`)
+      return
+    }
     function* cellsGenerator() {
       for (let i = 0; i < size; ++i) {
         yield convertWithImport(valueGetter(i), edit)
@@ -273,6 +300,7 @@ export function useTableNewArgument(
     maxWidth: 40,
     headerComponentParams: {
       type: 'newColumn',
+      enabled: mayAddNewColumn(),
       newColumnRequested: () => {
         const edit = graph.startEdit()
         fixColumns(edit)
@@ -382,7 +410,9 @@ export function useTableNewArgument(
         }
       }
     }
-    rows.push({ index: rows.length, cells: {} })
+    if (mayAddNewRow()) {
+      rows.push({ index: rows.length, cells: {} })
+    }
     return rows
   })
 
@@ -434,7 +464,7 @@ export function useTableNewArgument(
   }
 
   function pasteFromClipboard(data: string[][], focusedCell: { rowIndex: number; colId: string }) {
-    if (data.length === 0) return
+    if (data.length === 0) return { rows: 0, columns: 0 }
     const edit = graph.startEdit()
     const focusedColIndex =
       findIndexOpt(columns.value, ({ id }) => id === focusedCell.colId) ?? columns.value.length
@@ -446,6 +476,9 @@ export function useTableNewArgument(
     }
     const pastedRowsEnd = focusedCell.rowIndex + data.length
     const pastedColsEnd = focusedColIndex + data[0]!.length
+    // First we assume we'll paste all data. If not, these vars will be updated.
+    let actuallyPastedRowsEnd = pastedRowsEnd
+    let actuallyPastedColsEnd = pastedColsEnd
 
     // Set data in existing cells.
     for (
@@ -467,11 +500,20 @@ export function useTableNewArgument(
     // Extend the table if necessary.
     const newRowCount = Math.max(pastedRowsEnd, rowCount.value)
     for (let i = rowCount.value; i < newRowCount; ++i) {
+      if (!mayAddNewRow(i)) {
+        actuallyPastedRowsEnd = i
+        break
+      }
+
       addRow(edit, (_colId, index) => newValueGetter(i, index))
     }
     const newColCount = Math.max(pastedColsEnd, columns.value.length)
     let modifiedColumnsAst: Ast.Vector | undefined
     for (let i = columns.value.length; i < newColCount; ++i) {
+      if (!mayAddNewColumn(newRowCount, i)) {
+        actuallyPastedColsEnd = i
+        break
+      }
       modifiedColumnsAst = addColumn(
         edit,
         `${DEFAULT_COLUMN_PREFIX}${i + 1}`,
@@ -481,7 +523,10 @@ export function useTableNewArgument(
       )
     }
     onUpdate({ edit })
-    return
+    return {
+      rows: actuallyPastedRowsEnd - focusedCell.rowIndex,
+      columns: actuallyPastedColsEnd - focusedColIndex,
+    }
   }
 
   return {
@@ -513,6 +558,8 @@ export function useTableNewArgument(
      * If the pasted data are to be placed outside current table, the table is extended.
      * @param data the clipboard data, as retrieved in `processDataFromClipboard`.
      * @param focusedCell the currently focused cell: will become the left-top cell of pasted data.
+     * @returns number of actually pasted rows and columns; may be smaller than `data` size in case
+     * it would exceed {@link CELLS_LIMIT}.
      */
     pasteFromClipboard,
   }